Update to general structure discussed during the meeting

+++ src/interfaces/models/PointInterface.java
@@ -3,6 +3,6 @@ package interfaces.models;
 public interface PointInterface {
     Float getX();
     Float getY();
-    Integer distance(PointInterface p);
+    Float distance(PointInterface p);
     Boolean equals(PointInterface p);
 }
